refactor(components): tidy structure, fix casing/typos, drop dead code
Naming/consistency:
- Rename lowercase `form/` -> `Form/` to match the PascalCase folder
  convention (only ProfileEditForm imported it).
- Fix typo folder `ModalContet/` -> `ModalContent/`.

Dead code:
- Delete root `Features.tsx` (shadowed; active import is Features/Features).
- Delete `ChangePasswordSucsess.tsx` (only a commented-out import) and drop
  that comment in ChangePassword.

Group page/feature-local components out of the components root (shared
primitives like Icon/Pulse/GoBack/ModalWindow intentionally stay at root):
- Landing sections -> `Landing/` (Appointment, Discount, Faq, HeroImage).
- DoctorPage-only -> `DoctorPage/` (EmptyCalendar, SlotUnavailableModal).
- `MobileMenu` -> `Header/`, `ButtonLogout` -> `Dashboard/`.
ModalSignUp and ForgotPassword stay at root (each shared by two features).

Import paths updated accordingly; relative image/Icon paths in moved
Landing files adjusted one level deeper.
