+++ src/components/DoctorPage/DoctorPage.tsx
@@ -9,7 +9,7 @@ import { Card, CardBody } from "@heroui/react";
 import AboutTab from "./AboutTab";
 import WorkTab from "./WorkTab";
 import EducationTab from "./EducationTab";
-import SlotUnavailableModal from "@/components/SlotUnavailableModal"
+import SlotUnavailableModal from "./SlotUnavailableModal"
 import clsx from "clsx";
 import DocReviews from "./DocReviews";
 import BookingCalendar from "@/components/BookingCalendar/BookingCalendar";
@@ -25,7 +25,7 @@ import dayjs from "dayjs";
 import { useBookingFlow } from "@/hooks/booking/useBookingFlow";
 import { Pulse } from "@/components/Pulse";
 import { useScheduleSlots } from "@/hooks/vets/useScheduleSlots";
-import EmptyCalendar from "@/components/EmptyCalendar";
+import EmptyCalendar from "./EmptyCalendar";
 import { useProfile } from "@/hooks/owners/useProfile";
 
 type Props = {
